+++ packages/ente-api/src/entries/entries.service.ts
@@ -91,9 +91,6 @@ export class EntriesService {
     paginationInfo: PaginationInformation
   ): Promise<Validation<FindAllEntriesFailure, EntryDto[]>> {
     switch (requestingUser.role) {
-      case Roles.ADMIN:
-        return Success(await this.entryRepo.findAll(paginationInfo));
-
       case Roles.MANAGER:
         const user = (await requestingUser.getDto()).some();
         return Success(
@@ -127,7 +124,7 @@ export class EntriesService {
   ) {
     const r = await this._findAll(requestingUser, paginationInfo);
     return r.map(entries =>
-      entries.map(e => EntriesService.blackenDto(e, requestingUser.role))
+      entries.map(e => EntriesService.blackenDto(e, requestingUser))
     );
   }
 
@@ -144,9 +141,6 @@ export class EntriesService {
       () => Fail(FindEntryFailure.EntryNotFound),
       e => {
         switch (requestingUser.role) {
-          case Roles.ADMIN:
-            return Success(e);
-
           case Roles.STUDENT:
             const belongsStudent = e.student.id === requestingUser.id;
             return belongsStudent
@@ -177,7 +171,7 @@ export class EntriesService {
 
   public async findOne(id: string, requestingUser: RequestContextUser) {
     const r = await this._findOne(id, requestingUser);
-    return r.map(e => EntriesService.blackenDto(e, requestingUser.role));
+    return r.map(e => EntriesService.blackenDto(e, requestingUser));
   }
 
   private async getCreateEntryDtoValidator() {
@@ -286,7 +280,7 @@ export class EntriesService {
     requestingUser: RequestContextUser
   ) {
     const r = await this._create(entry, requestingUser);
-    return r.map(e => EntriesService.blackenDto(e, requestingUser.role));
+    return r.map(e => EntriesService.blackenDto(e, requestingUser));
   }
 
   private async _patch(
@@ -396,18 +390,14 @@ export class EntriesService {
     requestingUser: RequestContextUser
   ) {
     const r = await this._patch(id, patch, requestingUser);
-    return r.map(r => EntriesService.blackenDto(r, requestingUser.role));
+    return r.map(r => EntriesService.blackenDto(r, requestingUser));
   }
 
   private async _delete(
     id: string,
     requestingUser: RequestContextUser
   ): Promise<Validation<DeleteEntryFailure, EntryDto>> {
-    if (
-      [Roles.PARENT, Roles.STUDENT, Roles.TEACHER, Roles.ADMIN].includes(
-        requestingUser.role
-      )
-    ) {
+    if (requestingUser.role !== Roles.MANAGER) {
       return Fail(DeleteEntryFailure.ForbiddenForRole);
     }
 
@@ -442,7 +432,7 @@ export class EntriesService {
 
   public async delete(id: string, requestingUser: RequestContextUser) {
     const r = await this._delete(id, requestingUser);
-    return r.map(e => EntriesService.blackenDto(e, requestingUser.role));
+    return r.map(e => EntriesService.blackenDto(e, requestingUser));
   }
 
   async sendNotification(entryId: string) {
@@ -480,13 +470,16 @@ export class EntriesService {
     return `${baseUrl}/entries/${entry.id}`;
   }
 
-  static blackenDto(entry: EntryDto, role: Roles): BlackedEntryDto {
+  static blackenDto(
+    entry: EntryDto,
+    requestingUser: RequestContextUser
+  ): BlackedEntryDto {
     const { student, slots } = entry;
 
     return {
       ...entry,
-      student: UsersService.blackenDto(student, role),
-      slots: slots.map(s => SlotsService.blackenDto(s, role))
+      student: UsersService.blackenDto(student, requestingUser),
+      slots: slots.map(s => SlotsService.blackenDto(s, requestingUser))
     };
   }
 }