+++ packages/ente-api/src/helpers/request-context.ts
@@ -15,6 +15,10 @@ export interface WithRole {
   role: Roles;
 }
 
+export interface WithAdminState {
+  isAdmin: boolean;
+}
+
 export interface WithUsername {
   username: string;
 }
@@ -28,6 +32,7 @@ export interface RequestContextUser
     WithChildrenIds,
     WithRole,
     WithUsername,
+    WithAdminState,
     WithDisplayname {
   getDto: () => Promise<Maybe<UserDto>>;
 }
@@ -36,7 +41,9 @@ export type RequestContext<T = {}> = T & {
   user: RequestContextUser;
 };
 
-export const Ctx = createParamDecorator((_, req: Request): RequestContext => {
-  const user = req.user as RequestContextUser;
-  return { user };
-});
+export const Ctx = createParamDecorator(
+  (_, req: Request): RequestContext => {
+    const user = req.user as RequestContextUser;
+    return { user };
+  }
+);