+++ packages/ente-types/src/dtos/patch-user.dto.ts
@@ -41,4 +41,8 @@ export class PatchUserDto {
   @IsOptional()
   @IsBoolean()
   subscribedToWeeklySummary?: boolean;
+
+  @IsOptional()
+  @IsBoolean()
+  isAdmin?: boolean;
 }